Fix serve dependencies
