+++ python/ray/serve/examples/echo_split.py
@@ -2,12 +2,24 @@
 Example of traffic splitting. We will first use echo:v1. Then v1 and v2
 will split the incoming traffic evenly.
 """
+import json
 import time
+
+from pygments import formatters, highlight, lexers
 
 import requests
 
 from ray import serve
-from ray.serve.utils import pformat_color_json
+
+
+def pformat_color_json(d):
+    """Use pygments to pretty format and colorize dictionary"""
+    formatted_json = json.dumps(d, sort_keys=True, indent=4)
+
+    colorful_json = highlight(formatted_json, lexers.JsonLexer(),
+                              formatters.TerminalFormatter())
+
+    return colorful_json
 
 
 def echo_v1(_):
